+++ docs/chapter10.md
@@ -1,3 +1,9 @@
 # React と TypeScript の実践演習
 `npm install styled-components`  
 `npm install @types/styled-components`  
+
+1. 1個のコンポーネントに実装
+2. コンポーネント化できそうな部分を分割
+3. ロジック分離できそうな部分をカスタムフック化
+
+と段階を踏んで開発すると 分かりやすい
